tests: replace db tests with new_dv tests
